front_new/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
pydantic==2.6.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
supabase==2.3.4
python-dotenv==1.0.1
aiohttp==3.9.3
python-telegram-bot==20.8
asyncpg==0.29.0